refactor: 지원확인페이지 타이틀박스 컴포넌트 적용
