+++ src/components/ApplyResult/ResultImage.js
@@ -1,5 +1,7 @@
 import styled from 'styled-components';
 
+import { ReactComponent as AbleFindSvg } from '../../assets/images/icon_find_able.svg';
+import { ReactComponent as DisableFindSvg } from '../../assets/images/icon_find_disable.svg';
 import { ReactComponent as FailFindImg } from '../../assets/images/icon_find_fail.svg';
 import { ReactComponent as SuccessFindImg } from '../../assets/images/icon_find_success.svg';
 
@@ -9,6 +11,10 @@ const ResultImage = ({ result }) => {
       return <SuccessFindIcon />;
     case 'FAIL':
       return <FailFindIcon />;
+    case 'FOCUS':
+      return <AbleFindIcon />;
+    case 'NONFOCUS':
+      return <DisableFindIcon />;
     default:
       return <></>;
   }
@@ -26,6 +32,28 @@ const SuccessFindIcon = styled(SuccessFindImg)`
 `;
 
 const FailFindIcon = styled(FailFindImg)`
+  @media ${({ theme }) => theme.devices.TABLET} {
+    width: 30px;
+    height: 64px;
+  }
+  @media ${({ theme }) => theme.devices.DESKTOP} {
+    width: 38px;
+    height: 80px;
+  }
+`;
+
+const AbleFindIcon = styled(AbleFindSvg)`
+  @media ${({ theme }) => theme.devices.TABLET} {
+    width: 30px;
+    height: 64px;
+  }
+  @media ${({ theme }) => theme.devices.DESKTOP} {
+    width: 38px;
+    height: 80px;
+  }
+`;
+
+const DisableFindIcon = styled(DisableFindSvg)`
   @media ${({ theme }) => theme.devices.TABLET} {
     width: 30px;
     height: 64px;
